[#224] refactor - AVCapturePhotoCaptureDelegate Extension 분리
